Revert "Refactor shaker API and move BioShake to shaking module (#905)"
This reverts commit 8adab4d13f9daea759d069e20a465c72e6246330.
