+++ pylabrobot/heating_shaking/inheco/thermoshake_backend.py
@@ -1,5 +1,3 @@
-import warnings
-
 from pylabrobot.heating_shaking.backend import HeaterShakerBackend
 from pylabrobot.temperature_controlling.inheco.temperature_controller import (
   InhecoTemperatureControllerBackend,
@@ -16,7 +14,7 @@ class InhecoThermoshakeBackend(InhecoTemperatureControllerBackend, HeaterShakerB
     await self.stop_shaking()
     await super().stop()
 
-  async def _start_shaking_command(self):
+  async def start_shaking(self):
     """Start shaking the device at the speed set by `set_shaker_speed`"""
 
     return await self.interface.send_command(f"{self.index}ASE1")
@@ -53,7 +51,7 @@ class InhecoThermoshakeBackend(InhecoTemperatureControllerBackend, HeaterShakerB
 
     return await self.interface.send_command(f"1SSS{shape}")
 
-  async def start_shaking(self, speed: float, shape: int = 0):
+  async def shake(self, speed: float, shape: int = 0):
     """Shake the shaker at the given speed
 
     Args:
@@ -62,16 +60,7 @@ class InhecoThermoshakeBackend(InhecoTemperatureControllerBackend, HeaterShakerB
 
     await self.set_shaker_speed(speed=speed)
     await self.set_shaker_shape(shape=shape)
-    await self._start_shaking_command()
-
-  async def shake(self, speed: float, shape: int = 0):
-    """Deprecated alias for ``start_shaking``."""
-    warnings.warn(
-      "InhecoThermoshakeBackend.shake() is deprecated. Use start_shaking() instead.",
-      DeprecationWarning,
-      stacklevel=2,
-    )
-    await self.start_shaking(speed=speed, shape=shape)
+    await self.start_shaking()
 
   @property
   def supports_locking(self) -> bool:
